+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## 4.4.0 - 2023-10-04
+### Added
+- Queue manager
+
+### Fixed 
+- Download translations triggered only after all of them are done
+
 ## 4.3.0 - 2023-10-04
 ### Added
 - Queue each translation file transfer separately